MRD-5816 Change version numbers to pre-4.2 TEST
I changed all of the likely "4.1" strings in the code, but this doesn't appear to be testable until Jenkins does a build tho.  As near as I can tell when the build is packaging the jar file it adds a version file and a timestamp file but just building the app in eclipse or intellij doesn't do that stuff.