MRD-5311 Threat View: Summary Target Ratings row is no longer in sync with upper table when changing Threat/Target column width
The problem seemed to be that a label was being added in the lower left that spanned 2 columns, so resizing the second column caused weirdness when the widths were trying to be synchronized.  I broke the label up into 2 pieces, one for each column, with the first one just being blank.