MRD-5457 Menu items under mac are wrapped with HTML tags
There appears to be issues with the current use of html in menus, it caused this err as well MRD-5830, because of this I just stripped out the markup from the item text.  Doesn't appear anything special was being done with the html so it doesn't look like any new bugs will be introduced by this.